+++ src/main/java/org/servantscode/ministry/db/RoleDB.java
@@ -1,6 +1,9 @@
 package org.servantscode.ministry.db;
 
 import org.servantscode.commons.db.DBAccess;
+import org.servantscode.commons.search.QueryBuilder;
+import org.servantscode.commons.search.SearchParser;
+import org.servantscode.commons.security.OrganizationContext;
 import org.servantscode.ministry.MinistryRole;
 
 import java.sql.*;
@@ -12,12 +15,17 @@ import static org.servantscode.commons.StringUtils.isEmpty;
 
 public class RoleDB extends DBAccess {
 
-    public int getCount(int ministryId, String search) {
-        String sql = format("SELECT count(1) from ministry_roles WHERE ministry_id=?%s", optionalWhereClause(search));
-        try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement(sql)){
+    private SearchParser<MinistryRole> searchParser;
 
-            stmt.setInt(1, ministryId);
+    public RoleDB() {
+        searchParser = new SearchParser<>(MinistryRole.class);
+    }
+
+    public int getCount(int ministryId, String search) {
+        QueryBuilder query = count().from("ministry_roles").search(searchParser.parse(search)).inOrg();
+//        String sql = format("SELECT count(1) from ministry_roles WHERE ministry_id=?%s", optionalWhereClause(search));
+        try ( Connection conn = getConnection();
+              PreparedStatement stmt = query.prepareStatement(conn)){
 
             try(ResultSet rs = stmt.executeQuery()) {
                 return rs.next() ? rs.getInt(1) : 0;
@@ -28,13 +36,11 @@ public class RoleDB extends DBAccess {
     }
 
     public List<MinistryRole> getRoles(int ministryId, String search, String sortField, int start, int count) {
-        String sql = format("SELECT * FROM ministry_roles WHERE ministry_id=?%s ORDER BY %s LIMIT ? OFFSET ?", optionalWhereClause(search), sortField);
+        QueryBuilder query = selectAll().from("ministry_roles").where("ministry_id=?", ministryId).search(searchParser.parse(search)).inOrg()
+                .sort(sortField).limit(count).offset(start);
+//        String sql = format("SELECT * FROM ministry_roles WHERE ministry_id=?%s ORDER BY %s LIMIT ? OFFSET ?", optionalWhereClause(search), sortField);
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement(sql)
-        ) {
-            stmt.setInt(1, ministryId);
-            stmt.setInt(2, count);
-            stmt.setInt(3, start);
+              PreparedStatement stmt = query.prepareStatement(conn)) {
 
             return processResults(stmt);
         } catch (SQLException e) {
@@ -43,10 +49,9 @@ public class RoleDB extends DBAccess {
     }
 
     public MinistryRole getRole(int id) {
-        try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("SELECT * FROM ministry_roles WHERE id=?")
-        ) {
-            stmt.setInt(1, id);
+        QueryBuilder query = selectAll().from("ministry_roles").withId(id).inOrg();
+        try(Connection conn = getConnection();
+              PreparedStatement stmt = query.prepareStatement(conn) ) {
 
             List<MinistryRole> results = processResults(stmt);
             return results.isEmpty()? null: results.get(0);
@@ -57,13 +62,14 @@ public class RoleDB extends DBAccess {
 
     public void create(MinistryRole role) {
         try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement("INSERT INTO ministry_roles(name, ministry_id, contact, leader) values (?,?,?,?)", Statement.RETURN_GENERATED_KEYS)
+             PreparedStatement stmt = conn.prepareStatement("INSERT INTO ministry_roles(name, ministry_id, contact, leader, org_id) values (?,?,?,?,?)", Statement.RETURN_GENERATED_KEYS)
         ){
 
             stmt.setString(1, role.getName());
             stmt.setInt(2, role.getMinistryId());
             stmt.setBoolean(3, role.isContact());
             stmt.setBoolean(4, role.isLeader());
+            stmt.setInt(5, OrganizationContext.orgId());
 
             if(stmt.executeUpdate() == 0) {
                 throw new RuntimeException("Could not create role: " + role.getName());
@@ -80,7 +86,7 @@ public class RoleDB extends DBAccess {
 
     public void update(MinistryRole role) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("UPDATE ministry_roles SET name=?, ministry_id=?, contact=?, leader=? WHERE id=?")
+              PreparedStatement stmt = conn.prepareStatement("UPDATE ministry_roles SET name=?, ministry_id=?, contact=?, leader=? WHERE id=? AND org_id=?")
         ){
 
             stmt.setString(1, role.getName());
@@ -88,6 +94,7 @@ public class RoleDB extends DBAccess {
             stmt.setBoolean(3, role.isContact());
             stmt.setBoolean(4, role.isLeader());
             stmt.setInt(5, role.getId());
+            stmt.setInt(6, OrganizationContext.orgId());
 
             if(stmt.executeUpdate() == 0)
                 throw new RuntimeException("Could not update role: " + role.getName());
@@ -99,10 +106,11 @@ public class RoleDB extends DBAccess {
 
     public boolean delete(MinistryRole role) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("DELETE FROM ministry_roles WHERE id=?")
+              PreparedStatement stmt = conn.prepareStatement("DELETE FROM ministry_roles WHERE id=? AND org_id=?")
         ){
 
             stmt.setInt(1, role.getId());
+            stmt.setInt(2, OrganizationContext.orgId());
 
             return stmt.executeUpdate() != 0;
         } catch (SQLException e) {
